+++ src/components/app/app.jsx
@@ -10,7 +10,7 @@ import AppHeader from '../app-header/app-header'
 import BurgerIngredients from '../burger-ingredients/burger-ingredients'
 import BurgerConstructor from '../burger-constructor/burger-constructor'
 import Modal from '../modal/modal'
-import IngredientDetail from '../modal-ingredient/ingredient-detail'
+import IngredientDetail from '../ingredient-detail/ingredient-detail'
 import ModalOrderInfo from '../modal/modal-order-info/modal-order-info'
 import { LoginPage } from '../../pages/user/login'
 import { RegisterPage } from '../../pages/user/register'
