+++ src/components/ingredient-detail/ingredient-detail.jsx
