+++ src/components/ingredient-detail/ingredient-detail.module.css
